+++ ahead/Ridge2/Ridge2Regressor.py
@@ -1,48 +1,11 @@
-import os
 import numpy as np
-import pandas as pd
-
-from rpy2.robjects.packages import importr
-from rpy2.robjects import FloatVector, numpy2ri
-import rpy2.robjects as robjects
-import rpy2.robjects.packages as rpackages
-from datetime import datetime
-from rpy2.robjects.vectors import StrVector
-
+import rpy2.robjects.conversion as cv
+from rpy2.robjects import default_converter, FloatVector, ListVector, numpy2ri, r
+from .. import config
 from ..utils import multivariate as mv
 from ..utils import unimultivariate as umv
 
-required_packages = ["ahead"]  # list of required R packages
-
-if all(rpackages.isinstalled(x) for x in required_packages):
-    check_packages = True  # True if packages are already installed
-else:
-    check_packages = False  # False if packages are not installed
-
-if check_packages == False:  # Not installed? Then install.
-
-    packages_to_install = [
-        x for x in required_packages if not rpackages.isinstalled(x)
-    ]
-
-    if len(packages_to_install) > 0:
-        base = importr("base")
-        utils = importr("utils")
-        base.options(
-            repos=base.c(
-                techtonique="https://techtonique.r-universe.dev",
-                CRAN="https://cloud.r-project.org",
-            )
-        )
-        utils.install_packages(StrVector(packages_to_install))
-        check_packages = True
-
-base = importr("base")
-stats = importr("stats")
-ahead = importr("ahead")
-
-
-class Ridge2Regressor(object):
+class Ridge2Regressor():
     """Random Vector functional link network model with 2 regularization parameters
 
     Parameters:
@@ -81,11 +44,23 @@ class Ridge2Regressor(object):
 
         type_pi: a string;
             Type of prediction interval (currently "gaussian",
-            or "bootstrap")
+            or "bootstrap" or "blockbootstrap")
+        
+        block_length: an integer
+            length of block for multivariate circular block bootstrap (`type_pi == blockbootstrap`)
 
         B: an integer;
-            Number of bootstrap replications for `type_pi == bootstrap`
+            Number of bootstrap replications for `type_pi == bootstrap` or `type_pi == blockbootstrap` 
+        
+        type_aggregation: a string;
+            Type of aggregation, ONLY for bootstrapping; either "mean" or "median"
 
+        centers: an integer;
+            Number of clusters for \code{type_clustering}
+
+        type_clustering: a string;
+            "kmeans" (K-Means clustering) or "hclust" (Hierarchical clustering) 
+            
         cl: an integer; 
             The number of clusters for parallel execution (done in R), for `type_pi == bootstrap`
 
@@ -153,21 +128,25 @@ class Ridge2Regressor(object):
 
     def __init__(
         self,
-        h=5,
-        level=95,
-        lags=1,
-        nb_hidden=5,
-        nodes_sim="sobol",
-        activation="relu",
-        a=0.01,
-        lambda_1=0.1,
-        lambda_2=0.1,
-        dropout=0,
-        type_pi="gaussian",
-        B=100,
-        cl=1,
-        date_formatting="original",
-        seed=123,
+        h = 5,
+        level = 95,
+        lags = 1,
+        nb_hidden = 5,
+        nodes_sim = "sobol",
+        activation = "relu",
+        a = 0.01,
+        lambda_1 = 0.1,
+        lambda_2 = 0.1,
+        dropout = 0,
+        type_pi = "gaussian",
+        block_length = 3, # can be NULL, but in R (use 0 in R instead of NULL for v0.7.0)
+        B = 100,
+        type_aggregation = "mean",
+        centers = 2,
+        type_clustering = "kmeans",
+        cl = 1,
+        date_formatting ="original",
+        seed =123,
     ):
 
         self.h = h
@@ -181,27 +160,37 @@ class Ridge2Regressor(object):
         self.lambda_2 = lambda_2
         self.dropout = dropout
         self.type_pi = type_pi
+        self.block_length = block_length
         self.B = B
+        self.type_aggregation = type_aggregation
+        self.centers = centers # can be NULL, but in R (use 0 in R instead of NULL for v0.7.0)
+        self.type_clustering = type_clustering
         self.cl = cl
         self.date_formatting = date_formatting
         self.seed = seed
+        self.input_df = None
 
+        self.fcast_ = None
         self.averages_ = None
         self.ranges_ = None
         self.output_dates_ = []
         self.mean_ = None
         self.lower_ = None
         self.upper_ = None
-        self.result_df_s_ = None
+        self.result_dfs_ = None
         self.sims_ = None
+        self.xreg_ = None
 
-    def forecast(self, df):
+    def forecast(self, df, xreg = None):
         """Forecasting method from `Ridge2Regressor` class
 
         Parameters:
 
             df: a data frame;
                 a data frame containing the input time series (see example)
+            
+            xreg: a numpy array or a data frame;
+                external regressors
 
         """
 
@@ -217,8 +206,52 @@ class Ridge2Regressor(object):
         # obtain time series forecast -----
 
         y = mv.compute_y_mts(self.input_df, frequency)
-        self.fcast_ = ahead.ridge2f(
+
+        if xreg is None: 
+
+            with cv.localconverter(default_converter + config.NONE_CONVERTER):
+                self.fcast_ = config.AHEAD_PACKAGE.ridge2f(
+                y,
+                xreg = xreg,
+                h=self.h,
+                level=self.level,
+                lags=self.lags,
+                nb_hidden=self.nb_hidden,
+                nodes_sim=self.nodes_sim,
+                activ=self.activation,
+                a=self.a,
+                lambda_1=self.lambda_1,
+                lambda_2=self.lambda_2,
+                dropout=self.dropout,
+                type_pi=self.type_pi,
+                block_length=self.block_length, # can be NULL, but in R (use 0 in R instead of NULL for v0.7.0)
+                B=self.B,
+                type_aggregation = self.type_aggregation,
+                centers = self.centers,  # can be NULL, but in R (use 0 in R instead of NULL for v0.7.0)
+                type_clustering = self.type_clustering,
+                cl=self.cl,
+                seed=self.seed,
+            ) 
+        
+        else: # xreg is not None:  
+       
+            try:
+                self.xreg_ = xreg.values  
+            except: 
+                self.xreg_ = config.DEEP_COPY(xreg) 
+
+            is_matrix_xreg = (len(self.xreg_.shape) > 1)
+
+            numpy2ri.activate()            
+
+            xreg_ = r.matrix(FloatVector(self.xreg_.flatten()), 
+                        byrow = True, nrow = self.xreg_.shape[0], 
+                        ncol = self.xreg_.shape[1]) if is_matrix_xreg else r.matrix(FloatVector(self.xreg_.flatten()), 
+                        byrow=True, nrow=self.xreg_.shape[0], ncol=1)
+
+            self.fcast_ = config.AHEAD_PACKAGE.ridge2f(
             y,
+            xreg = xreg_,  
             h=self.h,
             level=self.level,
             lags=self.lags,
@@ -230,10 +263,14 @@ class Ridge2Regressor(object):
             lambda_2=self.lambda_2,
             dropout=self.dropout,
             type_pi=self.type_pi,
+            block_length=self.block_length, # can be NULL, but in R (use 0 in R instead of NULL for v0.7.0)
             B=self.B,
+            type_aggregation = self.type_aggregation,
+            centers = self.centers,  # can be NULL, but in R (use 0 in R instead of NULL for v0.7.0)
+            type_clustering = self.type_clustering,
             cl=self.cl,
             seed=self.seed,
-        )
+        )                     
 
         # result -----
 
@@ -258,7 +295,7 @@ class Ridge2Regressor(object):
             for i in range(n_series)
         )
 
-        if self.type_pi == "bootstrap":
+        if self.type_pi == "bootstrap" or self.type_pi == "blockbootstrap":
             self.sims_ = tuple(
                 np.asarray(self.fcast_.rx2["sims"][i]) for i in range(self.B)
             )