+++ ahead/VAR/VAR.py
@@ -1,46 +1,8 @@
-import os
 import numpy as np
-import pandas as pd
-
-from rpy2.robjects.packages import importr
-from rpy2.robjects import FloatVector, numpy2ri
-import rpy2.robjects as robjects
-import rpy2.robjects.packages as rpackages
-from datetime import datetime
-from rpy2.robjects.vectors import StrVector
+from .. import config
 
 from ..utils import multivariate as mv
 from ..utils import unimultivariate as umv
-
-required_packages = ["ahead"]  # list of required R packages
-
-if all(rpackages.isinstalled(x) for x in required_packages):
-    check_packages = True  # True if packages are already installed
-else:
-    check_packages = False  # False if packages are not installed
-
-if check_packages == False:  # Not installed? Then install.
-
-    packages_to_install = [
-        x for x in required_packages if not rpackages.isinstalled(x)
-    ]
-
-    if len(packages_to_install) > 0:
-        base = importr("base")
-        utils = importr("utils")
-        base.options(
-            repos=base.c(
-                techtonique="https://techtonique.r-universe.dev",
-                CRAN="https://cloud.r-project.org",
-            )
-        )
-        utils.install_packages(StrVector(packages_to_install))
-        check_packages = True
-
-base = importr("base")
-stats = importr("stats")
-ahead = importr("ahead")
-
 
 class VAR(object):
     """Vector AutoRegressive model
@@ -132,14 +94,16 @@ class VAR(object):
         self.lags = lags
         self.type_VAR = type_VAR
         self.date_formatting = date_formatting
+        self.input_df = None
 
+        self.fcast_ = None
         self.averages_ = None
         self.ranges_ = None
         self.output_dates_ = []
         self.mean_ = None
         self.lower_ = None
         self.upper_ = None
-        self.result_df_s_ = None
+        self.result_dfs_ = None
 
     def forecast(self, df):
         """Forecasting method from `VAR` class
@@ -163,7 +127,7 @@ class VAR(object):
         # obtain time series forecast -----
 
         y = mv.compute_y_mts(self.input_df, frequency)
-        self.fcast_ = ahead.varf(
+        self.fcast_ = config.AHEAD_PACKAGE.varf(
             y,
             h=self.h,
             level=self.level,