+++ ahead/utils/unimultivariate.py
@@ -1,7 +1,6 @@
 import numpy as np
 import pandas as pd
-
-from fuzzywuzzy import process
+from difflib import SequenceMatcher
 
 def compute_output_dates(df, horizon):
 
@@ -20,7 +19,6 @@ def compute_output_dates(df, horizon):
 
     return output_dates, frequency
 
-
 def compute_result_df(averages, ranges):
     pred_mean = pd.Series(dict(averages)).to_frame("mean")
     pred_ci = pd.DataFrame(
@@ -28,6 +26,9 @@ def compute_result_df(averages, ranges):
     ).set_index("date")
     return pd.concat([pred_mean, pred_ci], axis=1)
 
+def get_closest_str(input_str, list_choices):
+    scores = np.asarray([SequenceMatcher(None, a = input_str, b = elt).ratio() for elt in list_choices])  
+    return list_choices[np.where(scores == np.max(scores))[0][0]]
 
 def get_frequency(input_str):
     # https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html#offset-aliases
@@ -68,7 +69,9 @@ def get_frequency(input_str):
         return frequency_choices[input_str]
 
     except: 
-        
-        closest_str=process.extractOne(input_str, list(frequency_choices.keys()))[0]
+
+        list_frequency_choices = list(frequency_choices.keys())  
+
+        closest_str = get_closest_str(input_str=input_str, list_choices=list_frequency_choices)
 
         return frequency_choices[closest_str]