+++ examples/ridge2regressor.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pandas as pd
 from ahead import Ridge2Regressor
+from ahead  import plot
 from time import time
-
 
 # Forecasting horizon
 h = 5
@@ -38,6 +38,7 @@ print(d1.ranges_[1])
 print("\n")
 print(d1.ranges_[2])
 print("\n")
+plot(d1, selected_series="series1")
 
 print("Example 2 -----")
 d2 = Ridge2Regressor(h = h, date_formatting = "original")
@@ -57,7 +58,8 @@ print(d2.ranges_[1])
 print("\n")
 print(d2.ranges_[2])
 print("\n")
-
+plot(d2, selected_series="series1")
+print("\n")
 
 print("Example 3 -----")
 
@@ -91,3 +93,101 @@ print("\n mean, lower, upper as numpy arrays: \n")
 print(d3.mean_)
 print(d3.lower_)
 print(d3.upper_)
+print("\n")
+plot(d3, selected_series="series1")
+print("\n")
+
+print("Example 4 -----")
+
+d4 = Ridge2Regressor(h = h, date_formatting = "original", 
+type_pi="bootstrap", B=3, seed=1)
+
+# Data frame containing the time series 
+dataset = {
+ 'date' : ['2001-01-01', '2001-02-01', '2001-03-01', '2001-04-01', '2001-05-01'],
+ 'series1' : [34, 30, 35.6, 33.3, 38.1],    
+ 'series2' : [4, 5.5, 5.6, 6.3, 5.1],
+ 'series3' : [100, 100.5, 100.6, 100.2, 100.1]}
+df = pd.DataFrame(dataset).set_index('date')
+
+xreg = np.asarray(range(1, df.shape[0] + 1), dtype='float64')
+
+print(f"\n xreg: {xreg} \n")
+
+start = time()
+d4.forecast(df, xreg = xreg)
+print(f"Elapsed: {time()-start} \n")
+
+print(d4.fcast_.rx2['mean'])
+print(d4.averages_[1])
+print(np.asarray(d4.fcast_.rx2['mean']))
+
+print(d4.fcast_.rx2['sims'][0])
+res = np.asarray(d4.fcast_.rx2['sims'][1])
+print(res)
+print(res.shape)
+
+print("\n result_dfs_: \n")
+print(d4.result_dfs_)
+
+print("\n sims_: \n")
+print(d4.sims_)
+
+print("\n output_dates_: \n")
+print(d4.output_dates_)
+
+print("\n mean, lower, upper as numpy arrays: \n")
+print(d4.mean_)
+print(d4.lower_)
+print(d4.upper_)
+print("\n")
+plot(d4, selected_series="series1")
+print("\n")
+
+# print("Example 5 -----")
+
+# d5 = Ridge2Regressor(h = h, date_formatting = "original", 
+# type_pi="bootstrap", B=3, seed=1)
+
+# # Data frame containing the time series 
+# dataset = {
+#  'date' : ['2001-01-01', '2001-02-01', '2001-03-01', '2001-04-01', '2001-05-01'],
+#  'series1' : [34, 30, 35.6, 33.3, 38.1],    
+#  'series2' : [4, 5.5, 5.6, 6.3, 5.1],
+#  'series3' : [100, 100.5, 100.6, 100.2, 100.1]}
+# df = pd.DataFrame(dataset).set_index('date')
+
+# xreg = np.asarray(range(1, df.shape[0] + 1), dtype='float64')
+
+# print(f"\n xreg: {xreg} \n")
+
+# start = time()
+# d5.forecast(df, xreg = xreg)
+# print(f"Elapsed: {time()-start} \n")
+
+# print(d5.fcast_.rx2['mean'])
+# print(d5.averages_[1])
+# print(np.asarray(d5.fcast_.rx2['mean']))
+
+# print(d5.fcast_.rx2['sims'][0])
+# res = np.asarray(d5.fcast_.rx2['sims'][1])
+# print(res)
+# print(res.shape)
+
+# print("\n result_dfs_: \n")
+# print(d5.result_dfs_)
+
+# print("\n sims_: \n")
+# print(d5.sims_)
+
+# print("\n output_dates_: \n")
+# print(d5.output_dates_)
+
+# print("\n mean, lower, upper as numpy arrays: \n")
+# print(d5.mean_)
+# print(d5.lower_)
+# print(d5.upper_)
+# print("\n")
+# plot(d5, selected_series="series1")
+# print("\n")
+